requirements.txt
Django>=4.0.1
django-cors-headers==3.14.0
drf-spectacular>=0.22.1
djangorestframework>=3.13.1
psycopg2>=2.9.3
Pillow>=9.1.0
django-model-utils
whitenoise
uwsgi>=2.0.20